+++ examples/wrapper-components/react-javascript/src/App.js
@@ -19,8 +19,10 @@ import Tag from './components/Tag/Tag';
 import Switch from './components/Switch/Switch';
 import Card from './components/Card/Card';
 import Dropdown from './components/Dropdown/Dropdown';
+import Tooltip from './components/Tooltip/Tooltip';
 import SingleSelect from './components/SingleSelect/SingleSelect';
 import MultiSelect from './components/MultiSelect/MultiSelect';
+import BasicTable from './components/BasicTable/BasicTable';
 
 
 function App() {
@@ -114,6 +116,13 @@ function App() {
 
       <h2>Dropdown</h2>
       <Dropdown />
+      <br />
+
+      <h2>Tooltip</h2>
+      <Tooltip />
+
+      <h2>Basic Table</h2>
+      <BasicTable />
       <br />
 
     </div>